Add crypto trading (Upbit), regime-aware strategies, and fix critical bugs
Crypto market support:
- Upbit broker client (market buy/sell, balance, positions)
- Upbit price provider (candles, current price)
- Crypto symbol universes (top10, top30)
- Regime-aware meta-strategy (bull/sideways/bear detection)
- 5 crypto strategies: range-trading, volatility-breakout, volume-spike,
  rsi-contrarian, crypto-meta
- Intraday ORB (Opening Range Breakout) trading
- Capital tracker for crypto with JSON persistence
- Web UI: crypto market toggle, KRW formatting

Bug fixes:
- Fix sell quantity mismatch (Upbit fee deduction causes insufficient_funds)
- Add getActualQuantity() to query real broker balance before selling
- Add sell fail counter to prevent infinite retry loops
- Fix TradingDaysSince() to use date-based comparison (not time-based)
- Add CalendarDaysSince() for crypto (24/7, no weekday filter)
- Fix commission rate: 0.05% for crypto vs 0.25% for stocks
- Fix intraday trades not updating CapitalTracker
- Make wakeMonitor async with 5s timeout to prevent blocking

+++ internal/daemon/tracker.go
@@ -30,7 +30,7 @@ type TradeLog struct {
 	Timestamp   time.Time `json:"timestamp"`
 	Symbol      string    `json:"symbol"`
 	Side        string    `json:"side"` // "BUY" or "SELL"
-	Quantity    int       `json:"quantity"`
+	Quantity    float64   `json:"quantity"`
 	Price       float64   `json:"price"`
 	Amount      float64   `json:"amount"`
 	Commission  float64   `json:"commission"` // 수수료
@@ -325,7 +325,7 @@ TIME
 	if len(s.Trades) > 0 {
 		report += "TRADES\n------\n"
 		for i, trade := range s.Trades {
-			report += fmt.Sprintf("  %d. [%s] %s %s x%d @ $%.2f = $%.2f (%s)\n",
+			report += fmt.Sprintf("  %d. [%s] %s %s x%.4g @ $%.2f = $%.2f (%s)\n",
 				i+1,
 				trade.Timestamp.Format("15:04:05"),
 				trade.Side,
